call the hideEU function after content load on every page
