+++ src/javascript/app/base/footer.js
@@ -1,6 +1,5 @@
 const BinarySocket = require('./socket');
 const Client       = require('../base/client');
-const hideEU       = require('../common/common_functions').hideEU;
 const isEuCountry  = require('../common/country_base').isEuCountry;
 const State        = require('../../_common/storage').State;
 
@@ -16,7 +15,6 @@ const Footer = (() => {
             } else {
                 showWarning(isEuCountry());
             }
-            hideEU();
         });
     };
