+++ src/javascript/app/pages/cashier/cashier.js
@@ -1,7 +1,6 @@
 const Client           = require('../../base/client');
 const Header           = require('../../base/header');
 const BinarySocket     = require('../../base/socket');
-const hideEU           = require('../../common/common_functions').hideEU;
 const isCryptocurrency = require('../../common/currency').isCryptocurrency;
 const getElementById   = require('../../../_common/common_functions').getElementById;
 const paramsHash       = require('../../../_common/url').paramsHash;
@@ -84,7 +83,6 @@ const Cashier = (() => {
         PaymentMethods: {
             onLoad: () => {
                 showContent();
-                hideEU();
             },
         },
     };
